+++ src/fs.rs
@@ -1,6 +1,6 @@
-use crate::entry::{EntryKind, EntryStat};
+use crate::entry::EntryStat;
 use std::path::Path;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 use std::{fmt::Debug, path::PathBuf};
 
 use dashmap::DashMap;
@@ -9,7 +9,6 @@ use std::time::Duration;
 
 #[derive(Debug, Default)]
 pub struct CachedFS {
-    mutex: Mutex<()>,
     entries: DashMap<PathBuf, Arc<FileEntry>>,
 }
 
@@ -20,26 +19,6 @@ pub struct FileEntry {
 }
 
 impl CachedFS {
-    pub fn stat(&self, path: &Path) -> std::io::Result<EntryStat> {
-        let stat = if let Ok(meta) = std::fs::metadata(path) {
-            let kind = if meta.is_file() {
-                EntryKind::File
-            } else if meta.is_dir() {
-                EntryKind::Dir
-            } else {
-                EntryKind::Unknown
-            };
-            let mtime = Some(meta.modified()?);
-            EntryStat { kind, mtime }
-        } else {
-            EntryStat {
-                kind: EntryKind::NonExist,
-                mtime: None,
-            }
-        };
-        Ok(stat)
-    }
-
     pub fn read_file(&self, path: &Path, file_stat: &EntryStat) -> std::io::Result<String> {
         if let Some(cached) = self.entries.get(path) {
             // check cache